Bugfix/Dev: Make CAAML6 files compliant with Niviz
Niviz does not accept the 'unknown' value for Roughness. Hence,
roughness is not written in CAAML6 files unless the presence of other
data that require the presence of this element. We nevertheless
ensure compliance with CAAML6 spec in case of different data provided
but no roughness thus the files may not be readable by niviz.

Bugfixes in the writting of grain size (nan values incorrectly treated).